sentenceLabellingTool/labellertool/requirements.txt
astroid==2.4.2
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
colorama==0.4.3
dnspython==1.16.0
docx2txt==0.8
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
nltk==3.5
numpy==1.19.0
pandas==1.0.5
Pillow==7.2.0
pycparser==2.20
pylint==2.5.3
PyMuPDF==1.17.2
pytesseract==0.3.4
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
tqdm==4.47.0
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
